cifs: Fix inability to write files >2GB to SMB2/3 shares
When connecting to SMB2/3 shares, maximum file size is set to non-LFS
maximum in superblock. This is due to cap_large_files bit being
different for SMB1 and SMB2/3 (where it is just an internal flag that is
not negotiated and the SMB1 one corresponds to multichannel capability,
so maybe LFS works correctly if server sends 0x08 flag) while
capabilities are checked always for the SMB1 bit in cifs_read_super().

The patch fixes this by checking for the correct bit according to the
protocol version.

CC: Stable <stable@kernel.org>
Signed-off-by: Jan Klos <honza.klos@gmail.com>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <smfrench@gmail.com>